Move evaluateTarget, evaluateTokens into separate module
